+++ lib/project_types/node/commands/serve.rb
@@ -2,20 +2,18 @@
 module Node
   class Command
     class Serve < ShopifyCLI::SubCommand
+      include ShopifyCLI::CommandOptions::CommandServeOptions
+
       prerequisite_task ensure_project_type: :node
       prerequisite_task :ensure_env, :ensure_dev_store
 
-      options do |parser, flags|
-        parser.on("--host=HOST") do |h|
-          flags[:host] = h.gsub('"', "")
-        end
-        parser.on("--port=PORT") { |port| flags[:port] = port }
-      end
+      parse_host_option
+      parse_port_option
 
       def call(*)
         project = ShopifyCLI::Project.current
-        url = options.flags[:host] || ShopifyCLI::Tunnel.start(@ctx)
-        @ctx.abort(@ctx.message("node.serve.error.host_must_be_https")) if url.match(/^https/i).nil?
+        tunnel_port = port.to_s
+        url = host || ShopifyCLI::Tunnel.start(@ctx, port: tunnel_port)
         project.env.update(@ctx, :host, url)
         ShopifyCLI::Tasks::UpdateDashboardURLS.call(
           @ctx,
@@ -30,16 +28,9 @@ module Node
 
         CLI::UI::Frame.open(@ctx.message("node.serve.running_server")) do
           env = project.env.to_h
-          env["PORT"] = port.to_s
+          env["PORT"] = tunnel_port
           @ctx.system("npm run dev", env: env)
         end
-      end
-
-      def port
-        return ShopifyCLI::Tunnel::PORT.to_s unless options.flags.key?(:port)
-        port = options.flags[:port].to_i
-        @ctx.abort(@ctx.message("node.serve.error.invalid_port", options.flags[:port])) unless port > 0
-        port
       end
 
       def self.help