+++ lib/project_types/php/messages/messages.rb
@@ -133,11 +133,8 @@ module PHP
           extended_help: <<~HELP,
             {{bold:Options:}}
               {{cyan:--host=HOST}}: Bypass running tunnel and use custom host. HOST must be HTTPS url.
+              {{cyan:--port=PORT}}: Use custom port.
             HELP
-
-          error: {
-            host_must_be_https: "HOST must be a HTTPS url.",
-          },
 
           open_info: <<~MESSAGE,
             {{*}} To install and start using your app, open this URL in your browser: