+++ lib/project_types/rails/commands/serve.rb
@@ -2,19 +2,18 @@
 module Rails
   class Command
     class Serve < ShopifyCLI::SubCommand
+      include ShopifyCLI::CommandOptions::CommandServeOptions
+
       prerequisite_task ensure_project_type: :rails
       prerequisite_task :ensure_env, :ensure_dev_store
 
-      options do |parser, flags|
-        parser.on("--host=HOST") do |h|
-          flags[:host] = h.gsub('"', "")
-        end
-      end
+      parse_host_option
+      parse_port_option
 
       def call(*)
         project = ShopifyCLI::Project.current
-        url = options.flags[:host] || ShopifyCLI::Tunnel.start(@ctx)
-        @ctx.abort(@ctx.message("rails.serve.error.host_must_be_https")) if url.match(/^https/i).nil?
+        tunnel_port = port.to_s
+        url = host || ShopifyCLI::Tunnel.start(@ctx, port: tunnel_port)
         project.env.update(@ctx, :host, url)
         ShopifyCLI::Tasks::UpdateDashboardURLS.call(
           @ctx,
@@ -30,7 +29,7 @@ module Rails
         CLI::UI::Frame.open(@ctx.message("rails.serve.running_server")) do
           env = ShopifyCLI::Project.current.env.to_h
           env.delete("HOST")
-          env["PORT"] = ShopifyCLI::Tunnel::PORT.to_s
+          env["PORT"] = tunnel_port
           env["GEM_PATH"] = Gem.gem_path(@ctx)
           if @ctx.windows?
             @ctx.system("ruby bin\\rails server", env: env)