+++ lib/project_types/rails/messages/messages.rb
@@ -170,11 +170,8 @@ module Rails
           extended_help: <<~HELP,
             {{bold:Options:}}
               {{cyan:--host=HOST}}: Bypass running tunnel and use custom host. HOST must be HTTPS url.
+              {{cyan:--port=PORT}}: Use custom port.
           HELP
-
-          error: {
-            host_must_be_https: "{{red:HOST must be a HTTPS url.}}",
-          },
 
           open_info: <<~MESSAGE,
             {{*}} To install and start using your app, open this URL in your browser: