+++ lib/shopify_cli/messages/messages.rb
@@ -14,6 +14,18 @@ module ShopifyCLI
         },
       },
       core: {
+        app: {
+          serve: {
+            error: {
+              invalid_port: "%s is not a valid port.",
+              host_must_be_https: "HOST must be a HTTPS url.",
+            },
+          },
+          help: <<~HELP,
+          Create and manage embedded apps
+            Usage: {{command:%s app [ rails | node | php ] }}
+          HELP
+        },
         error_reporting: {
           unhandled_error: {
             message: "{{x}} {{red:An unexpected error occured.}}",