+++ test/project_types/php/commands/serve_test.rb
@@ -26,7 +26,7 @@ module PHP
           "artisan",
           "serve",
           "--port",
-          "3000",
+          "8081",
           env: {
             "SHOPIFY_API_KEY" => "mykey",
             "SHOPIFY_API_SECRET" => "mysecretkey",
@@ -59,7 +59,7 @@ module PHP
           "artisan",
           "serve",
           "--port",
-          "3000",
+          "8081",
           env: {
             "SHOPIFY_API_KEY" => "mykey",
             "SHOPIFY_API_SECRET" => "mysecretkey",
@@ -79,8 +79,17 @@ module PHP
         run_cmd("php serve")
       end
 
-      def test_server_command_with_invalid_host_url
-        ShopifyCLI::Tunnel.stubs(:start).returns("garbage://example.com")
+      def test_update_env_with_host
+        ShopifyCLI::Tunnel.expects(:start).never
+        ShopifyCLI::Tasks::UpdateDashboardURLS.expects(:call)
+        ShopifyCLI::Resources::EnvFile.any_instance.expects(:update).with(
+          @context, :host, "https://example-foo.com"
+        )
+        run_cmd('php serve --host="https://example-foo.com"')
+      end
+
+      def test_server_command_when_invalid_host_passed
+        invalid_host = "garbage://example.com"
         ShopifyCLI::Tasks::UpdateDashboardURLS.expects(:call).never
         ShopifyCLI::Resources::EnvFile.any_instance.expects(:update).never
         ShopifyCLI::ProcessSupervision.expects(:stop).never
@@ -91,7 +100,7 @@ module PHP
           "artisan",
           "serve",
           "--port",
-          "3000",
+          "8081",
           env: {
             "SHOPIFY_API_KEY" => "mykey",
             "SHOPIFY_API_SECRET" => "mysecretkey",
@@ -103,17 +112,56 @@ module PHP
         ).never
 
         assert_raises ShopifyCLI::Abort do
-          run_cmd("php serve")
+          run_cmd("php serve --host=#{invalid_host}")
         end
       end
 
-      def test_update_env_with_host
-        ShopifyCLI::Tunnel.expects(:start).never
+      def test_server_command_when_port_passed
+        ShopifyCLI::Tunnel.stubs(:start).returns("https://example.com")
         ShopifyCLI::Tasks::UpdateDashboardURLS.expects(:call)
-        ShopifyCLI::Resources::EnvFile.any_instance.expects(:update).with(
-          @context, :host, "https://example-foo.com"
+        ShopifyCLI::Resources::EnvFile.any_instance.expects(:update)
+        ShopifyCLI::ProcessSupervision.expects(:running?).with(:npm_watch).returns(false)
+        ShopifyCLI::ProcessSupervision.expects(:stop).never
+        ShopifyCLI::ProcessSupervision.expects(:start).with(:npm_watch, "npm run watch", force_spawn: true)
+
+        @context.expects(:system).with(
+          "php",
+          "artisan",
+          "serve",
+          "--port",
+          "5000",
+          env: {
+            "SHOPIFY_API_KEY" => "mykey",
+            "SHOPIFY_API_SECRET" => "mysecretkey",
+            "SHOP" => "my-test-shop.myshopify.com",
+            "SCOPES" => "read_products",
+            "HOST" => "https://example.com",
+            "DB_DATABASE" => "storage/db.sqlite",
+          }
         )
-        run_cmd('php serve --host="https://example-foo.com"')
+
+        @context.expects(:puts).with(
+          "\n" +
+          @context.message("php.serve.open_info", "https://example.com/login?shop=my-test-shop.myshopify.com") +
+          "\n"
+        )
+
+        run_cmd("php serve --port=5000")
+      end
+
+      def test_server_command_when_invalid_port_passed
+        invalid_port = "NOT_PORT"
+        ShopifyCLI::Tunnel.stubs(:start).returns("https://example.com")
+        ShopifyCLI::Tasks::UpdateDashboardURLS.expects(:call)
+        ShopifyCLI::Resources::EnvFile.any_instance.expects(:update)
+        ShopifyCLI::ProcessSupervision.expects(:running?).with(:npm_watch).returns(false)
+        ShopifyCLI::ProcessSupervision.expects(:stop).never
+        ShopifyCLI::ProcessSupervision.expects(:start).with(:npm_watch, "npm run watch", force_spawn: true)
+        @context.expects(:abort).with(
+          @context.message("core.app.serve.error.invalid_port", invalid_port)
+        )
+
+        run_cmd("php serve --port=#{invalid_port}")
       end
     end
   end